added error checking and label

+++ src/application/Clothes.java
@@ -3,17 +3,20 @@ package application;
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
+import javafx.scene.control.TextField;
 
 public class Clothes extends Item {
 	private SimpleStringProperty size;
 	private SimpleStringProperty material;
 
-	public Clothes(int id, String name, int cost, int salePrice, int amount, int totalCost, int totalSalePrice,
+	public Clothes(int id, String name, double cost, double salePrice, int amount, double totalCost, double totalSalePrice,
 			String size, String material) {
 		super(id, name, cost, salePrice, amount, totalCost, totalSalePrice);
 		this.size = new SimpleStringProperty(size);
 		this.material = new SimpleStringProperty(material);
 	}
+	//https://stackoverflow.com/questions/35569061/java-error-checking-for-text-field
+	
 
 	public String getSize() {
 		return size.get();
@@ -40,5 +43,7 @@ public class Clothes extends Item {
 	public void setMaterial(String material) {
 		this.material.set(material);
 	}
+	
+	
 
 }
